Patch RangerBasePlugin to forward authorization requests to a Rest endpoint instead of a central Ranger server.
This is most probably not workable in this state, but it was in a compiling state a while ago, so can hopefully be salvaged.

Most specifically I have not added anything to the pom to import gson dependencies - maven didn't complain, but that is probably just due to the changed code not having been pulled when building the actual authorizer.